Extract is_thrust_pointer into separate file to avoid nvrtc error
